fix: move subproject parser into a helper
